@@ -1671,8 +1671,8 @@ namespace Crystallography
 								pixelVertex[n++] = new PointD((tempY2numer1 + tempX * Numer2) * numer4, (tempX * Numer1 + tempY2numer3) * numer4);
 							}
 						}
-						PointD temp = pixelVertex[2]; pixelVertex[2] = pixelVertex[3]; pixelVertex[3] = temp;//���v���ɂȂ�悤��3�߂�4�߂����ւ���
-						double totalArea = Geometriy.GetPolygonalArea(pixelVertex);
+                        (pixelVertex[3], pixelVertex[2]) = (pixelVertex[2], pixelVertex[3]);
+                        double totalArea = Geometriy.GetPolygonalArea(pixelVertex);
 
 						//���߂�4���̓_����A�X�e�b�v�̎w���̏���A������ݒ�
 						List<double> angles = new List<double>();
